LayoutTests/transitions/lengthsize-transition-to-from-auto-expected.txt

CONSOLE MESSAGE: A transition for property background-size could not be found
CONSOLE MESSAGE: Failed to pause 'background-size' transition on element 'test2'
CONSOLE MESSAGE: A transition for property background-size could not be found
CONSOLE MESSAGE: Failed to pause 'background-size' transition on element 'test3'
PASS - "background-size" property for "test1" element at 1s saw something close to: 55,10
PASS - "background-size" property for "test2" element at 1s saw something close to: 10
PASS - "background-size" property for "test3" element at 1s saw something close to: 10,10
